requirements.txt
openai>=1.12.0
clickhouse-connect>=0.8.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.6
aiofiles>=23.2.1
gradio>=4.19.0
